product details page thone and login ui thone

+++ app/Http/Controllers/Dpanel/ProductController.php
@@ -150,21 +150,23 @@ class ProductController extends Controller
             $variant->save();
         }
 
-        foreach ($req->images as $key=>$image){
+        if ($req->images){
+            foreach ($req->images as $key=>$image){
 
-            if (isset($req->image_ids[$key])){
-                $productImage =  ProductImage::find($req->image_ids[$key]);
-            Storage::disk('public')->delete($productImage->path);
-                $productImage->path = $image->store('media','public');
-                $productImage->save();
+                if (isset($req->image_ids[$key])){
+                    $productImage =  ProductImage::find($req->image_ids[$key]);
+                    Storage::disk('public')->delete($productImage->path);
+                    $productImage->path = $image->store('media','public');
+                    $productImage->save();
 
-            }else{
-                $productImage = new ProductImage;
-                $productImage->product_id = $product->id;
-                $productImage->path = $image->store('media','public');
-                $productImage->save();
+                }else{
+                    $productImage = new ProductImage;
+                    $productImage->product_id = $product->id;
+                    $productImage->path = $image->store('media','public');
+                    $productImage->save();
+                }
+
             }
-
         }
 
         return redirect()->route('dpanel.product.index')->with('success','New Products added successfully');
